[TagEditor.py] Add internal tag editor
This module replaces all the various tag editor plugins that were created in the absence of a central and integrated tag editor.  This new code addresses bugs in the most commonly used tag editor plugin.  This code correctly returns the appropriate values for timer tag edits and does not corrupt movie meta files if their tag entries are edited.

This code offers integrated button HELP and uses the VirtualKeyBoard to edit the tag text for a more enhanced editing experience.

The code also defines and uses a single TagMaster class to manage and coordinate all tag related management functions.

Co-authored-by: jbleyel <jbleyel@gmx.net>