[TemplatedMultiContent.py] Improve assert statements
This change causes the assert statements to return an error message to the higher level code (skin.py) if the asserts are active and cause the code to abort.  With out the error text the try/except code in skin.py has no information about the error that caused the except to trigger.